给定时任务-录音文件删除，oss文件删除，放音文件同步，增加redis锁。

+++ app-backend/src/main/java/com/lsxy/app/backend/task/DeleteOssFileTask.java
@@ -1,6 +1,10 @@
 package com.lsxy.app.backend.task;
 
+import com.lsxy.framework.cache.exceptions.TransactionExecFailedException;
+import com.lsxy.framework.cache.manager.RedisCacheService;
 import com.lsxy.framework.config.SystemConfig;
+import com.lsxy.framework.core.utils.DateUtils;
+import com.lsxy.framework.core.utils.StringUtil;
 import com.lsxy.framework.oss.OSSService;
 import com.lsxy.yunhuni.api.config.service.LineGatewayService;
 import com.lsxy.yunhuni.api.file.model.VoiceFilePlay;
@@ -13,6 +17,7 @@ import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 import java.util.Map;
 
@@ -32,18 +37,56 @@ public class DeleteOssFileTask {
     private VoiceFilePlayService voiceFilePlaydService;
     @Autowired
     private LineGatewayService lineGatewayService;
+    @Autowired
+    private RedisCacheService redisCacheService;
     /**
      * 通知进行放音文件同步操作
      */
     @Scheduled(cron="2 0/30 * * * ?")
     public void startSync() {
-        if(logger.isDebugEnabled()){
-            logger.debug("启动放音文件/录音文件删除未成功补救方案--开始");
+        Date date=new Date();
+        String month = DateUtils.formatDate(date, "yyyy-MM-dd HH:mm");
+        String cacheKey = "scheduled_" + Thread.currentThread().getStackTrace()[1].getClassName() + "." + Thread.currentThread().getStackTrace()[1].getMethodName() + " " + month;
+        //执行互斥处理消息
+        String flagValue = redisCacheService.get( cacheKey);
+        if(StringUtil.isNotEmpty(flagValue)){
+            if(logger.isDebugEnabled()){
+                logger.debug("["+cacheKey+"]缓存中已被设置标记，该任务被"+flagValue+"处理了");
+            }
+        }else{
+            try {
+                if(logger.isDebugEnabled()){
+                    logger.debug("["+cacheKey+"]准备处理该任务:"+cacheKey);
+                }
+                redisCacheService.setTransactionFlag(cacheKey, SystemConfig.id,60*30);
+                String currentCacheValue = redisCacheService.get(cacheKey);
+                if(logger.isDebugEnabled()){
+                    logger.debug("["+cacheKey+"]当前cacheValue:"+currentCacheValue);
+                }
+                if(currentCacheValue.equals(SystemConfig.id)){
+                    if(logger.isDebugEnabled()){
+                        logger.debug("["+cacheKey+"]马上处理该任务："+cacheKey);
+                    }
+                    //执行语句
+                    if(logger.isDebugEnabled()){
+                        logger.debug("启动放音文件/录音文件删除未成功补救方案--开始");
+                    }
+                    deleteOssFile();
+                    if(logger.isDebugEnabled()){
+                        logger.debug("启动放音文件/录音文件删除未成功补救方案--结束");
+                    }
+                }else{
+                    if(logger.isDebugEnabled()){
+                        logger.debug("["+cacheKey+"]标记位不一致"+currentCacheValue+"  vs "+ SystemConfig.id);
+                    }
+                }
+            } catch (TransactionExecFailedException e) {
+                if(logger.isDebugEnabled()){
+                    logger.debug("["+cacheKey+"]设置标记位异常了，该任务被另一节点处理了");
+                }
+            }
         }
-        deleteOssFile();
-        if(logger.isDebugEnabled()){
-            logger.debug("启动放音文件/录音文件删除未成功补救方案--结束");
-        }
+
     }
     private void deleteOssFile(){
         deleteOssPlayFile();
